requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-dotenv==1.4.2
idna==3.4
Pillow==9.1.1
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.3.1
requests==2.31.0
sqlparse==0.4.2
typing_extensions==4.7.1
urllib3==2.0.3
